refactor & update configuration to contain custom query periods
